formatters -> eval_cache
Group all the evaluation cache functions together
